refactor: separate file for accepted commits
